+++ packages/core/src/internal/coreEffectRunners/cancelled.js
@@ -1,0 +1,3 @@
+export default function runCancelledEffect(env, effectPayload, cb, { mainTask }) {
+  cb(Boolean(mainTask._isCancelled))
+}
