+++ packages/core/src/internal/coreEffectRunners/flush.js
@@ -1,0 +1,3 @@
+export default function flush(env, channel, cb) {
+  channel.flush(cb)
+}
