+++ packages/core/src/internal/coreEffectRunners/getContext.js
@@ -1,0 +1,3 @@
+export default function getContext(env, prop, cb, { taskContext }) {
+  cb(taskContext[prop])
+}
