+++ packages/core/src/internal/coreEffectRunners/put.js
@@ -1,0 +1,26 @@
+import { asap } from '../scheduler'
+import { is } from '../utils'
+
+export default function put(env, { channel, action, resolve }, cb, { resolvePromise }) {
+  /**
+   Schedule the put in case another saga is holding a lock.
+   The put will be executed atomically. ie nested puts will execute after
+   this put has terminated.
+   **/
+  asap(() => {
+    let result
+    try {
+      result = (channel ? channel.put : env.dispatch)(action)
+    } catch (error) {
+      cb(error, true)
+      return
+    }
+
+    if (resolve && is.promise(result)) {
+      resolvePromise(result, cb)
+    } else {
+      cb(result)
+    }
+  })
+  // Put effects are non cancellables
+}
