+++ packages/core/src/internal/coreEffectRunners/setContext.js
@@ -1,0 +1,6 @@
+import { object } from '../utils'
+
+export default function setContext(env, props, cb, { taskContext }) {
+  object.assign(taskContext, props)
+  cb()
+}
