+++ packages/core/src/internal/coreEffectRunners/take.js
@@ -1,0 +1,25 @@
+import { isEnd } from '../channel'
+import { is } from '../utils'
+import { TERMINATE } from '../symbols'
+import matcher from '../matcher'
+
+export default function take(env, { channel = env.stdChannel, pattern, maybe }, cb) {
+  const takeCb = input => {
+    if (input instanceof Error) {
+      cb(input, true)
+      return
+    }
+    if (isEnd(input) && !maybe) {
+      cb(TERMINATE)
+      return
+    }
+    cb(input)
+  }
+  try {
+    channel.take(takeCb, is.notUndef(pattern) ? matcher(pattern) : null)
+  } catch (err) {
+    cb(err, true)
+    return
+  }
+  cb.cancel = takeCb.cancel
+}
